backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic-settings==2.1.0
elevenlabs==0.2.23
pytest==7.4.3
pytest-asyncio==0.23.2
groq==0.4.2
email-validator==2.1.0
